+++ components/charts/ChartLineView.tsx
@@ -1,4 +1,5 @@
 import { appColors } from "@/utils/colors";
+import { useEffect, useRef } from "react";
 import { Dimensions, ScrollView, Text, View } from "react-native";
 import { LineChart } from "react-native-chart-kit";
 
@@ -50,7 +51,20 @@ export const ChartLineView = ({
       strokeWidth: "3",
     },
   };
+    const scrollRef = useRef<ScrollView>(null);
 
+  useEffect(() => {
+    if (values.length > 0) {
+      setTimeout(() => {
+        scrollRef.current?.scrollTo({ x: 0, animated: false }); 
+        animateScrollToEnd();
+      }, 400);
+    }
+  }, [values]);
+  
+  const animateScrollToEnd = () => {
+    scrollRef.current?.scrollToEnd({ animated: true });
+  };
   if (values.length === 0) {
     return (
       <View className="flex items-center justify-center py-6">
@@ -69,6 +83,7 @@ export const ChartLineView = ({
         </Text>
       )}
       <ScrollView
+        ref={scrollRef}
         horizontal
         showsHorizontalScrollIndicator={false}
         style={{ borderRadius: 12, marginTop: 5 }}