+++ modules/goals/components/FastFilters.tsx
@@ -12,6 +12,8 @@ interface Props {
 
   sortByAssigned: boolean;
   setSortByAssigned: (value: boolean) => void;
+
+
 }
 
 export default function FastFilters({