+++ modules/goals/components/GoalsFilterModal.tsx
@@ -1,17 +1,29 @@
 import FilterModal from "@/components/ui/FilterModal";
 import { appColors } from "@/utils/colors";
 import { useEffect, useState } from "react";
-import { ActivityIndicator, Text, TouchableOpacity, View } from "react-native";
+import {
+  ActivityIndicator,
+  ScrollView,
+  Text,
+  TouchableOpacity,
+  View,
+} from "react-native";
 import { Seller } from "../types/Seller";
-
-
 
 interface GoalsFilterModalProps {
   visible: boolean;
   onClose: () => void;
-  onApply: (selectedSellers: string[]) => void;
+  onApply: (
+    selectedSellers: string[],
+    selectedCategory: string | undefined,
+    selectedUsedValue: string | undefined
+  ) => void;
   sellers: Seller[];
+  category: any[];
   selectedSellers: string[]; // selección inicial
+  selectedCategory: string | undefined; // selección inicial
+  usedValues: any[];
+  seletedUsedValue: string | undefined;
   hasPermission: boolean;
   loading: boolean;
 }
@@ -22,6 +34,10 @@ export default function GoalsFilterModal({
   onApply,
   sellers,
   selectedSellers,
+  category,
+  selectedCategory,
+  usedValues,
+  seletedUsedValue,
   hasPermission,
   loading,
 }: GoalsFilterModalProps) {
@@ -42,13 +58,29 @@ export default function GoalsFilterModal({
         : [...prev, co_ven]
     );
   };
+  const [internalCategory, setInternalCategory] = useState<string | undefined>(
+    selectedCategory
+  );
+  const [internalUsedValue, setInternalUsedValue] = useState<
+    string | undefined
+  >(seletedUsedValue);
+
+  useEffect(() => {
+    if (visible) {
+      setInternalSelected(selectedSellers || []);
+      setInternalCategory(selectedCategory);
+      setInternalUsedValue(seletedUsedValue);
+    }
+  }, [visible, selectedSellers, selectedCategory, seletedUsedValue]);
 
   const handleApply = () => {
-    onApply(internalSelected); // solo aquí pasamos la selección al padre
+    onApply(internalSelected, internalCategory, internalUsedValue);
   };
 
   const handleClean = () => {
     setInternalSelected([]);
+    setInternalCategory(undefined);
+    setInternalUsedValue(undefined);
   };
 
   return (
@@ -72,33 +104,106 @@ export default function GoalsFilterModal({
           </View>
         ) : (
           <>
-            <Text className="pb-2 text-md font-semibold text-foreground dark:text-dark-foreground">
-              Seleccione uno o más vendedores
+            <Text className="mb-1 font-medium text-mutedForeground dark:text-dark-mutedForeground">
+              Estado de uso
             </Text>
-   
-            <View className="flex-row flex-wrap gap-2 mb-3">
-              {sellers.map((seller) => (
+            <ScrollView
+              contentContainerClassName="p-2 gap-2 mb-3"
+              horizontal
+              showsHorizontalScrollIndicator={false}
+            >
+              {usedValues?.map((opt) => (
                 <TouchableOpacity
-                  key={seller.codven}
-                  onPress={() => handleSelect(seller.codven)}
+                  key={opt.value}
+                  onPress={() =>
+                    setInternalUsedValue((prev) =>
+                      prev === opt.value ? undefined : opt.value
+                    )
+                  }
                   className={`px-4 py-2 rounded-full border ${
-                    internalSelected.includes(seller.codven)
+                    internalUsedValue === opt.value
                       ? "bg-primary border-primary"
                       : "bg-transparent border-muted"
                   }`}
                 >
                   <Text
                     className={`text-sm ${
-                      internalSelected.includes(seller.codven)
+                      internalUsedValue === opt.value
                         ? "text-white"
                         : "text-foreground dark:text-dark-foreground"
                     }`}
                   >
-                    {seller.vendes}
+                    {opt.label}
                   </Text>
                 </TouchableOpacity>
               ))}
-            </View>
+            </ScrollView>
+            <Text className="mb-1 font-medium text-mutedForeground dark:text-dark-mutedForeground">
+              Categoría
+            </Text>
+            <ScrollView
+              contentContainerClassName=" p-2 gap-2 mb-3"
+              horizontal
+              showsHorizontalScrollIndicator={false}
+            >
+              {category?.map((opt) => (
+                <TouchableOpacity
+                  key={opt.codcat}
+                  onPress={() =>
+                    setInternalCategory((prev) =>
+                      prev === opt.codcat ? undefined : opt.codcat
+                    )
+                  }
+                  className={`px-4 py-2 rounded-full border ${
+                    internalCategory === opt.codcat
+                      ? "bg-primary border-primary"
+                      : "bg-transparent border-muted"
+                  }`}
+                >
+                  <Text
+                    className={`text-sm ${
+                      internalCategory === opt.codcat
+                        ? "text-white"
+                        : "text-foreground dark:text-dark-foreground"
+                    }`}
+                  >
+                    {opt.catdes?.trim()}
+                  </Text>
+                </TouchableOpacity>
+              ))}
+            </ScrollView>
+
+            {hasPermission && (
+              <>
+                <Text className="mb-1 font-medium text-mutedForeground dark:text-dark-mutedForeground">
+                  Seleccione uno o más vendedores
+                </Text>
+
+                <View className="flex-row flex-wrap gap-3 mb-3 pt-2">
+                  {sellers.map((seller) => (
+                    <TouchableOpacity
+                      key={seller.codven}
+                      onPress={() => handleSelect(seller.codven)}
+                      className={`px-4 py-2 rounded-full border ${
+                        internalSelected.includes(seller.codven)
+                          ? "bg-primary border-primary"
+                          : "bg-transparent border-muted"
+                      }`}
+                    >
+                      <Text
+                        className={`text-sm ${
+                          internalSelected.includes(seller.codven)
+                            ? "text-white"
+                            : "text-foreground dark:text-dark-foreground"
+                        }`}
+                      >
+                        {seller.vendes}
+                      </Text>
+                    </TouchableOpacity>
+                  ))}
+                </View>
+              </>
+            )}
           </>
         )}
       </View>