+++ modules/goals/hooks/useGoalsResumen.ts
@@ -1,7 +1,8 @@
 import { useRefreshControl } from "@/utils/userRefreshControl";
 import { useFocusEffect } from "expo-router";
 import { useCallback, useEffect, useMemo, useState } from "react";
-import { getGoals, getSellersGoals } from "../services/GoalsService";
+import { getCategorys, getGoals, getSellersGoals } from "../services/GoalsService";
+import { Category } from "../types/Category";
 import { Goals } from "../types/Goals";
 import { Seller } from "../types/Seller";
 
@@ -13,16 +14,25 @@ export function useGoalsResumen(searchText: string) {
   const [error, setError] = useState<string | null>(null);
   const [allGoals, setAllGoals] = useState<Goals[]>([]);
   const [sellers, setSellers] = useState<Seller[]>([]);
+  const [category, setCategorys] = useState<Category[]>([]);
 
   // Filters
   const [notUsed, setNotUsed] = useState<boolean>(false);
   const [sortByUsed, setSortByUsed] = useState<boolean>(false);
   const [sortByAssigned, setSortByAssigned] = useState<boolean>(false);
 
-  // Sellers selection
+  // Filters selection
   const [selectedSellers, setSelectedSellers] = useState<string[]>([]);
-  const [loadingSellers, setLoadingSellers] = useState(false);
+  const [selectedCategory, setSelectedCategory] = useState<string>();
+  const [loadingFilters, setLoadingFilters] = useState(false);
+  const [selectedUsedValue, setSelectedUsedValue] = useState<string>();
+  const usedValues = [
+    { label: "TODOS", value: "TODOS" },
+    { label: "USADOS", value: "USADOS" },
+    { label: "NO USADOS", value: "NO USADOS" },
+    { label: "CUMPLIDAS", value: "CUMPLIDAS" },
 
+  ];
   const { refreshing, canRefresh, cooldown, wrapRefresh, cleanup } = useRefreshControl(10);
 
   const [initialized, setInitialized] = useState(false);
@@ -46,16 +56,30 @@ export function useGoalsResumen(searchText: string) {
   }, []);
 
   /** Load sellers */
-  const loadSellers = useCallback(async () => {
-
-    setLoadingSellers(true);
+  const loadFilters = useCallback(async () => {
+    setLoadingFilters(true);
     try {
-      const result = await getSellersGoals();
-      setSellers(result.map((s) => ({ codven: s.codven, vendes: s.vendes.trim() })));
+      const [sellersResult, categoryResult] = await Promise.all([
+        getSellersGoals(),
+        getCategorys()
+      ]);
+
+      setSellers(
+        sellersResult.map((s) => ({
+          codven: s.codven,
+          vendes: s.vendes.trim(),
+        }))
+      );
+      setCategorys([
+        { codcat: "TODOS", catdes: "TODOS" },
+        ...categoryResult
+      ]);
+
+
     } catch (err) {
       console.error(err);
     } finally {
-      setLoadingSellers(false);
+      setLoadingFilters(false);
     }
   }, []);
 
@@ -83,8 +107,8 @@ export function useGoalsResumen(searchText: string) {
 
   useFocusEffect(
     useCallback(() => {
-     
-      loadSellers();
+
+      loadFilters();
       loadGoals(selectedSellers.length > 0 ? selectedSellers : undefined);
       setError(null);
       setInitialized(true);
@@ -120,10 +144,18 @@ export function useGoalsResumen(searchText: string) {
     }
 
     // Not used filter
-    if (notUsed) {
+    if (notUsed || selectedUsedValue==='NO USADOS') {
       filteredgoals = filteredgoals.filter((g) => g.utilizado < 1);
     }
+    if (selectedUsedValue ==='USADOS') {
+      filteredgoals = filteredgoals.filter((g) => g.utilizado > 0);
+    } else if (selectedUsedValue === 'CUMPLIDAS'){
+      filteredgoals = filteredgoals.filter((g) => g.utilizado===g.asignado);
+    }
 
+    if(selectedCategory && selectedCategory!=='TODOS'){
+      filteredgoals = filteredgoals.filter((g) => g.catdes.startsWith(selectedCategory));
+    }
     // Sorting
     if (sortByUsed) {
       filteredgoals.sort((a, b) => b.utilizado - a.utilizado);
@@ -131,11 +163,12 @@ export function useGoalsResumen(searchText: string) {
     if (sortByAssigned) {
       filteredgoals.sort((a, b) => b.asignado - a.asignado);
     }
+  
 
     setGoals(filteredgoals);
     setLoading(false)
 
-  }, [allGoals, searchText, notUsed, sortByUsed, sortByAssigned]);
+  }, [allGoals, searchText, notUsed, sortByUsed, sortByAssigned, selectedCategory, selectedUsedValue]);
 
   /** Summary */
   const resumen = useMemo(() => {
@@ -144,7 +177,7 @@ export function useGoalsResumen(searchText: string) {
     const totalDisponible = totalAsignada - totalUtilizado;
     const totalPercent = totalAsignada > 0 ? totalUtilizado / totalAsignada : 0;
     const totalArticles = goals.length;
-    const totalFilters = selectedSellers.length > 0 ? 1 : 0;
+    const totalFilters = (selectedSellers.length > 0 ? 1 : 0) + (selectedCategory != undefined ? 1 : 0) + (selectedUsedValue != undefined ? 1 : 0) ;
     return { totalAsignada, totalUtilizado, totalDisponible, totalPercent, totalArticles, totalFilters };
   }, [goals]);
 
@@ -154,8 +187,14 @@ export function useGoalsResumen(searchText: string) {
     sellers,
     selectedSellers,
     setSelectedSellers,
+    category,
+    selectedCategory,
+    setSelectedCategory,
+    setSelectedUsedValue,
+    selectedUsedValue,
+    usedValues,
     loading,
-    loadingSellers,
+    loadingFilters,
     error,
     ...resumen,
     notUsed,